Notifications/News_Notification/requirements.txt
pyarrow
requests
google-cloud-bigquery
flask
twilio